[refact] 주문 서비스 에러처리 고도화

+++ msa.order/src/main/java/com/msa/order/application/client/dto/request/ProductStockRequest.java
@@ -1,5 +1,5 @@
 package com.msa.order.application.client.dto.request;
 
-public record ProductStockRequest(int stock) {
+public record ProductStockRequest(Long stock) {
 
 }
